+++ src/configs/index.js
@@ -1,5 +1,8 @@
+require('dotenv').config();
+
 const config = {
-    DB_CONNECTION_STRING: process.env.MONGODB_URL,
-  };
-  
-  export default config;
+  DB_CONNECTION_STRING: process.env.MONGODB_URL,
+  PORT:process.env.PORT
+};
+
+module.exports = config;
